add caching

+++ src/main/java/com/example/userservice/service/impl/PaymentCardServiceImpl.java
@@ -1,8 +1,8 @@
 package com.example.userservice.service.impl;
 
-import com.example.userservice.dto.PaymentCardCreateDto;
-import com.example.userservice.dto.PaymentCardResponseDto;
-import com.example.userservice.dto.PaymentCardUpdateDto;
+import com.example.userservice.dto.paymentCard.PaymentCardCreateDto;
+import com.example.userservice.dto.paymentCard.PaymentCardResponseDto;
+import com.example.userservice.dto.paymentCard.PaymentCardUpdateDto;
 import com.example.userservice.dto.filter.PaymentCardFilter;
 import com.example.userservice.entity.PaymentCard;
 import com.example.userservice.entity.User;
@@ -14,6 +14,7 @@ import com.example.userservice.repository.PaymentCardRepository;
 import com.example.userservice.repository.UserRepository;
 import com.example.userservice.service.PaymentCardService;
 import com.example.userservice.specification.PaymentCardSpecification;
+import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -37,6 +38,7 @@ public class PaymentCardServiceImpl implements PaymentCardService {
 
   @Override
   @Transactional
+  @CacheEvict(value = "userWithCards", key = "#dto.userId()")
   public PaymentCardResponseDto create(PaymentCardCreateDto dto) {
     User user = userRepository.findById(dto.userId())
             .orElseThrow(() -> new UserNotFoundException(dto.userId()));
@@ -74,6 +76,7 @@ public class PaymentCardServiceImpl implements PaymentCardService {
 
   @Override
   @Transactional
+  @CacheEvict(value = "userWithCards", key = "#result.userId")
   public PaymentCardResponseDto update(Long id, PaymentCardUpdateDto dto) {
     PaymentCard card = cardRepository.findById(id)
             .orElseThrow(() -> new PaymentCardNotFoundException(id));
@@ -83,26 +86,37 @@ public class PaymentCardServiceImpl implements PaymentCardService {
 
   @Override
   @Transactional
-  public void activate(Long id) {
+  @CacheEvict(value = "userWithCards", key = "#result")
+  public Long activate(Long id) {
     if (!cardRepository.existsById(id)) {
       throw new PaymentCardNotFoundException(id);
     }
+    Long userId = cardRepository.findUserIdByCardId(id);
     cardRepository.updateActiveStatus(id, true);
+    return userId;
   }
 
   @Override
   @Transactional
-  public void deactivate(Long id) {
+  @CacheEvict(value = "userWithCards", key = "#result")
+  public Long deactivate(Long id) {
     if (!cardRepository.existsById(id)) {
       throw new PaymentCardNotFoundException(id);
     }
+    Long userId = cardRepository.findUserIdByCardId(id);
     cardRepository.updateActiveStatus(id, false);
+    return userId;
   }
 
   @Override
   @Transactional
-  public void delete(Long id) {
+  @CacheEvict(value = "userWithCards", key = "#result")
+  public Long delete(Long id) {
+    Long userId = cardRepository.findUserIdByCardId(id);
+    if (userId == null) {
+      throw new PaymentCardNotFoundException(id);
+    }
     cardRepository.deleteById(id);
+    return userId;
   }
-
 }
